Rename datasheet generator
