+++ scripts/generate-datasheets.js
@@ -1,5 +1,5 @@
 const fileHelper = require('./lib/file-helper');
-const parser = require("./datasheet-generator/parser");
+const generator = require("./datasheet-generator/datasheet-generator");
 
 const DATASHEETS_SOURCE_PATH = "../content/datasheets";
 const DATASHEETS_TARGET_PATH = "../build/datasheets";
@@ -10,7 +10,7 @@ const DATASHEETS_TARGET_PATH = "../build/datasheets";
     for(let filePath of datasheetFiles){
         console.log(`Generating datasheet for ${filePath} ...`);
         fileHelper.createDirectoryIfNecessary(DATASHEETS_TARGET_PATH)                
-        await parser.generatePDFFromMarkdown(filePath, DATASHEETS_TARGET_PATH);
+        await generator.generatePDFFromMarkdown(filePath, DATASHEETS_TARGET_PATH);
     }
     console.log("✅ %s Datasheets generated.", datasheetFiles.length);
     process.exit(0);
